requirements.txt
beautifulsoup4==4.12.3
networkx==2.8.8
tqdm==4.66.5
aiosqlite==0.20.0
aiofiles==24.1.0
uvloop==0.20.0
pymgclient==1.3.1
gqlalchemy==1.6.0
rapidfuzz==3.10.0
xxhash==3.5.0
cython==3.0.11
PyYAML==6.0.2
pandas==2.0.3
lxml==4.9.3
pytest==7.3.1
flake8==6.0.0
black==23.1.0
pytest-asyncio==0.24.0
pytest==8.3.3
pydantic
pydantic-settings
pytest-mock
pytest-cov
ruff